Init File: Activate New PVT Table Normalisation
This commit replaces the existing calls to

    Tables::addPVTO()
    Tables::addPVTG()
    Tables::addPVTW()

with the new umbrella member function

    Tables::addPVTTables()

This, in turn, activates table normalisation for all PV{D,T}{G,O},
PVCDO, and PVTW keywords and produces ECLIPSE compatibly sized TAB
vector representations for these data sources.

As a prerequisite for this work, to enable users such as ResInsight
to correctly interpret the table contents, we need to ensure that
the contents of LOGIHEAD reflect the data sources used to form TAB.
As a consequence, we now control the contents of INTEHEAD, LOGIHEAD,
and DOUBHEAD directly from EclipseIO.cpp rather than from LibECL
library function ecl_init_file_fwrite_header().

We reuse header functionality from the restart files.  This changes
the size of the *HEAD vectors in the init file:

    Vector    Old  New
    INTEHEAD   95  411
    LOGIHEAD   80  121
    DOUBHEAD    1  229

which matches ECLIPSE 2017.2.  Note that this is taking a bit of a
shortcut, because there are items in restart files that differ from
those of the init file.  However, for the items that affect tables,
the items are the same in both init and restart files.  We will need
to refine this at a later point.