Ecl Output: Add Support for Opening File at Specific Position
This commit extends the EclOutput constructor to support letting the
client specify a particular file position and a particular open
mode.  These extensions collectively prepare using EclOutput as the
basis for a restart file writer.  Passing mode=openmode::out is the
common case for a new restart file, while mode=openmode::app is the
common case for an existing restart file.  The write position is a
special extension, mostly to support a very special use case that is
exercised in the unit tests but typically not in a real simulation
run.

We implement the write position request in terms of Boost (and std)
Filesystem's resize_file() function which essentially just wraps the
POSIX function ::truncate().