RestartIO: Reimplement save() in Terms of EclOutput
This commit switches the file handling operations of the gateway
function RestartIO::save() to using the FileService component.
Briefly, this means replacing write_kw() function calls with calls
to EclOutput::write(), especially since the later natively handles
CharArrayNullTerm<8> data.

The most involved piece of logic in this commit concerns adding new
data to an existing unified restart file.  In this case we need to
pay attention to the file opening mode and, especially, to the
position at which we place the output pointer.  Consequently, the
new helper function 'existingRestartFile' uses class ERst's support
for identifying sequence numbers and translating those into file
positions, as well as class EclOutput's support for opening a file
at a particular write position.