RestartIO: Reimplement load() in Terms of ERst
This commit reimplements the private RestartFileView class in terms
of the ERst class.  The latter supports all operations we currently
need, although we can no longer use a 'const' object since loading
arrays mutates the ERst instance.

Most changes here are fairly mechanic and most of the new complexity
concerns how to identify a vector (keyword) of a particular element
type.  We make special concessions for INTEHEAD, since this vector
plays a central role in accessing the other data vectors.  The
mutability of RestartFileView is underscored by using shared_ptr<>
in the various helper facilities.