INIT File: Output Scaled End-Point Arrays for Saturation Functions
This commit introduces a new helper function,

    writeSatFuncScaling()

that will extract scaled end-point arrays like SWL, SGCR, SOWCR,
KRORW, and PCG and output these to the INIT file.  We support both
the drainage and imbibition keywords, but do not support the
directional versions of these keywords at this time.

Note that this first draft uses the keyword values directly from the
Eclipse3DProperties container.  Therefore, the output behaves as if
the input deck always specifies the 'FILLEPS' option.  Outputting
only the values that were specified in the input will be the subject
of future work.