Add Base Protocol for New Summary Parameter Evaluation
This commit introduces a new base class, Opm::SummaryParameter, that
formalises the system currently implemented in Summary.cpp .  In
particular, class SummaryParameter specifies an interface for
calculating values of summary parameters and storing those values as
updates in a SummaryState object.

Add a concrete implementation of the interface for block parameters
(e.g., BOSAT or BPR) and a set of unit tests for these parameter types.