Summary Parameter: Add Protocol For Evaluating Well Parameters
This commit introduces a new representation of well-related summary
parameters.  The class Opm::SummaryHelpers::WellParameter represents
a well parameter and will invoke an evaluation function in its
update member function.  We store that numerical value in the
SummaryState object through the 'update_well_parameter()' function.

We support pressure (i.e., BHP and THP), flow rates, rate ratios
(e.g., GOR, GLR, and WCT), and cumulative total parameter types.
The client must specify which parameter type is needed at
construction time and should additionally call 'validate' on the
parameter object.

The set of wells is trivial for well parameters so we do not need to
do any searching before invoking the evaluation function.  We only
need to check that the well is active on the current simulation
step.

Add a set of unit tests to exercise the WellParameter interface on a
set of different evaluation function, both with and without
well/group efficiency factors for cumulative totals.