Summary Parameter: Add Protocol For Evaluating Group Parameters
This commit introduces a new representation of well-related summary
parameters.  The class Opm::SummaryHelpers::GroupParameter
represents a group parameter and will invoke an evaluation function
in its update member function.  We store that numerical value in the
SummaryState object through the 'update_group_parameter()' function.
The set of wells involved in calculating the numerical value is derived
from a group tree traversal down from the group itself.  We implement
that traversal through a helper function, groupTreeTraversal, which
calls user-specified operations on groups (class Group2) and wells
(class Well2) as needed.  This traversal is also useful in deriving
aggregate efficiency factors for child groups and wells.

Similarly, the class Opm::SummaryHelpers::FieldParameter plays the same
role for the FIELD group.  The only difference between Group and FIELD
is that the latter uses SummaryState::update() and its set of wells is
more easily determined and we don't need a group tree traversal to do
so.

We support well counts, flow rates, rate ratios (e.g., GOR, GLR, and
WCT), and cumulative total parameter types.  The client must specify
which parameter type is needed at construction time and should
additionally call 'validate' on the parameter object.

Add a set of unit tests to exercise the Group- and FieldParameter
interfaces on a set of different evaluation function, both with and
without well/group efficiency factors for cumulative totals.