Summary Parameter: Add Protocol For Evaluating Segment Parameters
This commit introduces a new representation of segment-related
summary parameters.  The class Opm::SummaryHelpers::SegmentParameter
represents a segment parameter and will invoke an evaluation
function in its update member function.  We store that numerical
value in the SummaryState object through the 'update()' function,
using a key that combines the keyword, the well name, and the
segment ID number.  The set of wells involved in calculating the
numerical value is trivial.

We support surface flow rates of the oil, gas, and water components
in addition to the segment pressure (SPR).  The client must specify
the pertinent parameter type at object construction time.  We
additionally recommend calling 'validate' on the parameter object.

Add a set of unit tests to exercise the interfaces on a set of
different evaluation functions.