Summary Parameter: Add Protocol For Evaluating Connection Parameters
This commit introduces a new representation of well connection
summary parameters.  The class ConnectionParameter stores the
numerical value into the SummaryState object through the 'update()'
function.

We support flow rates and cumulative totals and, additionally, the
connection transmissibility factor (CTFAC) which also accounts for
the WPIMULT factor.

Add a set of unit tests to exercise the interfaces on a set of
different evaluation functions and connection transmissibility
factors.