Summary Parameter: Add Protocol For Global Process Parameters
This commit introduces a new representation of global "process" summary
parameters.  This parameter type includes TCPU, FPR, FOIP, TIMESTEP
and others such as number of linear and non-linear iterations.  The
values are typically extracted directly from the simulator (and
communicated through a map<string,double>).  The class stores the
numerical value into the SummaryState object through the 'update()'
function.

We support all parameter types that can be represented in the map
format and associated to a particular unit of measure.

Add a couple of unit tests to exercise the interfaces.