Prepare Restart File Support for Valves and ICDs
This commit identifies certain items in restart vectors ISEG and
RSEG that pertain to valves (keyword WSEGVALV) and spiral ICDs
(keyword WSEGSICD).  In particular, we identify segment type flag
and a couple of relevant values for this flag.  Also, relevant
parameters that describe the geometry and pressure drops of these
devices.