Address performance issue for DX/DY/DZ/TOPS grids.
The sumIdir() and sumJdir() methods were called inside loop
nests over i and j. Since the methods themselves were linear in
nx and ny, respectively, this was a quadratic algorithm. Performance
only becomes an issue for large (> 10k) values of NX or NY, which
was thought not to ever happen. This assumption was wrong.