Sat-Func Properties: Provide All Raw Endpoints to All Initializers
This is in preparation of fixing the initializers for the KR*R
properties.  The default values here generally depend on critical
and connate saturations of other phases extracted from the raw input
table values.  This commit therefore creates an aggregate structure

    RawTableEndPoints

that collects the connate gas and water saturations, the full set of
critical saturations, and the maximum gas and water saturations from
all input tables (meaning for all regions and drainage/imbibition
regime designators).  Each *Endpoint() function furthermore gets a
new parameter of this type to preserve a uniform signature.

Note that we currently calculate the complete set of these endpoints
for each call to satfunc::init(), but may choose to expose this
structure to the caller at a later time in order to reduce the
overhead of these repeated calculations.